[1차 과제 - Step3] 강지윤 (#18)
* add (domain) : create Subway

* test : add api test for post station

* add (controller) : create SubwayController

* add (service) : create SubwayService

* add (repository) : create SubwayRepository

* add (DTO) : create DTOs for post station

* feat : implement to create station

* test : add api test for get stations

* add (DTO) : create DTO for get stations

* feat : implement to get stations list

* test : add api test for delete station

* feat : implement to delete station by id

* refactor: move api test code

* rename : rename 'subway' to 'station'

* move : relocated station test code

* rename : rename 'subway' to 'station'

* feat : add duplicate name check in createStation

* rename (PostStationResponse)

* refactor : consolidate station api endpoints

* refactor : organize station-related DTOs

* chore : add Hibernate Validator dependency

* refactor : add input validation

* refactor : enhance encapsulation & remove Setter

* refactor : simplify code

* add : create StationStep

* refactor : enhance Readability and Maintainability

* feat : add method for creating station

* refactor : simplify code by extracting duplicates

* add : create ControllerAdvice

* refactor : simplify code in stream

* refactor : wrap return type with ResponseEntity

* feat : add exception handler

* test : add create station test with null value

* test : create LineTest

* refactor : add static field in StationStep

* add : create BaseResponseStatus

* feat : create BasseResponse

* feat : add BaseResponseEntity extends ResponseEntity

* feat : create DomainException

* feat : apply BaseResponseEntity on return type

* refactor : replace find method with exists method

* refactor : apply static factory methods & rename DTOs

* move : relocated station DTOs

* test : add api test for post line

* feat : add RestAssuredUtil for simplified HTTP interactions

POST요청, GET요청 구현

* create : add LineController

* feat : create Line entity

* create : add LineService

* create : add LineRepository

* feat : implement CreateLineResponse DTO

정적 팩토리 메서드 패턴으로 생성 구현

* feat : create LineRequest DTO with validation

* feat : implement to create line

* create : LineException

* test : add api test for get line by Id

* feat : implement ShowLineResponse DTO

정적 팩토리 메서드 패턴으로 생성 구현

* feat : implement to show line by Id

* feat : create LineFixture

* test : add api test for get line list

* feat : implement to show line list

* test : add api test for update line

* feat : implement to update line

* test : add api test case for update line

* test : add api test for delete line

* feat : implement to delete line

* refactor : add ResponseStatus

* refactor : optimize annotations in request DTO

필요한 어노테이션만 사용하도록 리팩토링
- @Setter, @AllArgsConstructor 는 사용하지 않음.
- 반면 @NoArgsConstructor 는 필요!
  Json 으로 변환하는 과정에서 필요함

* refactor : set field to private

* refactor : add Transactional annotation

* add Classist Mockist Jgraph

* test : create SectionTest

* test : add api test for add section to line

* add : create StationFixture

* refactor : change body type to Object

in RestAssuredUtil

* feat : implement basic method to add Section

* add : create Section Entity

* feat : replace Section List with Sections object

List<Section>을 일급컬렉션으로 관리

* feat : implement SectionService

* feat : implement section registration validation

* feat : create SectionRepository

* feat : implement bidirectional addition

* feat : implement ShowLineResponse with Sections

* refactor : seperate LineRequestDTO

* test : refactor update line

* refactor : change body type & add pathParam

* refactor : refactor update & create line

구간 개념 도입으로 인해 요구 사항이 변경됨

* test : add comparison logic for responseCode

* refactor : organize code using Fixture and Step

* test : add api test for delete section to line

* feat : implement api for delete section to line

* refactor : add validation in deleteLine

* refactor : make field private

* move (AcceptanceTest)

* refactor : seperate validate method

* refactor : change FetchType to LAZY

* fix : move methods from Service & Line to Sections

* remove (SectionService)

* fix : refactor delete section method

* refactor : remove StationService Dependency

* refactor : merge from() and fromSection()

* rename : renamed classes related to AcceptanceTest

---------

+++ src/main/java/kuit/subway/service/LineService.java
@@ -1,12 +1,18 @@
 package kuit.subway.service;
 
 import kuit.subway.domain.Line;
+import kuit.subway.domain.Section;
 import kuit.subway.domain.Station;
 import kuit.subway.repository.LineRepository;
-import kuit.subway.request.line.LineRequest;
-import kuit.subway.response.line.CreateLineResponse;
+import kuit.subway.repository.SectionRepository;
+import kuit.subway.repository.StationRepository;
+import kuit.subway.request.line.CreateLineRequest;
+import kuit.subway.request.line.UpdateLineRequest;
+import kuit.subway.request.section.DeleteSectionRequest;
+import kuit.subway.request.section.SectionRequest;
 import kuit.subway.response.line.ShowLineResponse;
 import kuit.subway.utils.exception.LineException;
+import kuit.subway.utils.exception.StationException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,26 +25,34 @@ import static kuit.subway.utils.BaseResponseStatus.*;
 @RequiredArgsConstructor
 public class LineService {
     private final LineRepository lineRepository;
-    private final StationService stationService;
+    private final SectionRepository sectionRepository;
+    private final StationRepository stationRepository;
 
     @Transactional
-    public CreateLineResponse createLine(LineRequest request) {
-        Station downStation = stationService.findById(request.getDownStationId());
-        Station upStation = stationService.findById(request.getUpStationId());
+    public ShowLineResponse createLine(CreateLineRequest request) {
+        Station downStation = stationRepository.findById(request.getDownStationId())
+                .orElseThrow(() -> new StationException(NONE_STATION));
+        Station upStation = stationRepository.findById(request.getUpStationId())
+                .orElseThrow(() -> new StationException(NONE_STATION));
 
-        checkDuplicateName(request.getName());
+        checkDuplicateName(request.getName(), null);
 
         Line line = Line.builder()
-                .downStation(downStation)
-                .upStation(upStation)
-                .distance(request.getDistance())
                 .name(request.getName())
                 .color(request.getColor())
                 .build();
 
         lineRepository.save(line);
 
-        return CreateLineResponse.from(line);
+        Section section = Section.builder()
+                .downStation(downStation)
+                .upStation(upStation)
+                .distance(request.getDistance())
+                .build();
+
+        section.addLine(line);
+        sectionRepository.save(section);
+        return ShowLineResponse.from(line);
     }
 
     @Transactional(readOnly = true)
@@ -57,21 +71,54 @@ public class LineService {
     }
 
     @Transactional
-    public ShowLineResponse updateLine(Long id, LineRequest request) {
+    public ShowLineResponse updateLine(Long id, UpdateLineRequest request) {
         Line line = findById(id);
-        Station downStation = stationService.findById(request.getDownStationId());
-        Station upStation = stationService.findById(request.getUpStationId());
+        checkDuplicateName(request.getName(), id);
+        checkDuplicateColor(request.getColor(), id);
 
-        checkDuplicateName(request.getName());
-
-        line.update(request, upStation, downStation);
+        line.update(request.getName(), request.getColor());
 
         return ShowLineResponse.from(line);
     }
 
     @Transactional
     public void deleteLine(Long id) {
+        if(!lineRepository.existsLineById(id))
+            throw new LineException(NONE_LINE);
         lineRepository.deleteById(id);
+    }
+
+    // requestBody를 바탕으로 Section을 생성한 후 Line에 추가한다.
+    @Transactional
+    public ShowLineResponse addSectionToLine(Long line_id, SectionRequest request) {
+        Line line = findById(line_id);
+
+        Station downStation = stationRepository.findById(request.getDownStationId())
+                .orElseThrow(() -> new StationException(NONE_STATION));
+
+        Station upStation = stationRepository.findById(request.getUpStationId())
+                .orElseThrow(() -> new StationException(NONE_STATION));
+
+        Section section = Section.builder()
+                .downStation(downStation)
+                .upStation(upStation)
+                .distance(request.getDistance())
+                .build();
+
+        section.addLine(line);
+        sectionRepository.save(section);
+
+        return ShowLineResponse.from(line);
+    }
+
+    @Transactional
+    public void deleteSection(Long line_id, DeleteSectionRequest request) {
+        Line line = findById(line_id);
+
+        Station station = stationRepository.findById(request.getStationId())
+                .orElseThrow(() -> new StationException(NONE_STATION));
+
+        line.deleteSectionByStation(station);
     }
 
     private Line findById(Long id) {
@@ -79,10 +126,20 @@ public class LineService {
                 .orElseThrow(() -> new LineException(NONE_LINE));
     }
 
-    private void checkDuplicateName(String name) {
-        if(lineRepository.existsLineByName(name))
-            throw new LineException(DUPLICATED_LINE);
+    private void checkDuplicateName(String name, Long line_id) {
+        if(!lineRepository.existsLineByName(name)) return;
+        if(line_id == null)
+            throw new LineException(DUPLICATED_LINENAME);
+        if(!lineRepository.findLineByName(name).getId().equals(line_id))
+            throw new LineException(DUPLICATED_LINENAME);
     }
 
+    private void checkDuplicateColor(String color, Long line_id) {
+        if(!lineRepository.existsLineByColor(color)) return;
+        if(line_id == null)
+            throw new LineException(DUPLICATED_LINECOLOR);
+        if(!lineRepository.findLineByColor(color).getId().equals(line_id))
+            throw new LineException(DUPLICATED_LINECOLOR);
+    }
 
 }
